backend/requirements.txt
# Django core
Django>=5.1,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0

# Production server
gunicorn>=22.0,<24.0

# Static files in production
whitenoise>=6.7,<7.0


